Refactor -> move simple resolver to separate mod
commit-id:3ba8d8f0
